[dep-link] Additional check to prevent linked pod cleaning and minor tweaks

+++ lib/cocoapods/local_pod.rb
@@ -1,8 +1,10 @@
 module Pod
   class LocalPod
-    attr_reader :specification
-    attr_reader :sandbox
-    
+
+    attr_reader   :specification
+    attr_reader   :sandbox
+    attr_accessor :link_path
+
     def initialize(specification, sandbox)
       @specification, @sandbox = specification, sandbox
     end
@@ -18,6 +20,8 @@ module Pod
     def to_s
       if specification.local?
         "#{specification} [LOCAL]"
+      elsif @link_path
+        "#{specification} -> #{@link_path}"
       else
         specification.to_s
       end
@@ -45,7 +49,7 @@ module Pod
     end
     
     def clean
-      clean_paths.each { |path| FileUtils.rm_rf(path) }
+      clean_paths.each { |path| FileUtils.rm_rf(path) } if !@link_path
     end
     
     def source_files
